requirements.txt
sounddevice 
webrtcvad 
numpy 
vosk 
faster-whisper

openai>=1.40.0
python-dotenv>=1.0.1
